Изменение : Заменил GetTickCount64() -> GetTickCount(), т.к. не поддерживается на WinXP;
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@3169 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/renderer/VideoRenderers/CpuUsage.h
@@ -45,7 +45,7 @@ private:
 	FILETIME m_ftPrevProcUser;
 
 	short m_nCpuUsage;
-	ULONGLONG m_dwLastRun;
+	DWORD m_dwLastRun;
 	
 	volatile LONG m_lRunCount;
 };
